// src/app/shared/models/admin-dashboard-data.ts
export interface AdminDashboardData {
  totalUsers: number;
  totalStudents: number;
  totalTrainers: number;
  activeUsers: number;
  totalCourses: number;
  activeCourses: number;
  totalBatches: number;
  activeBatches: number;
  totalQuizzes: number;
  activeQuizzes: number;
  todayAttendance: number;
  totalFeedback: number;
}